Implement goto statements
Note: There are currently zero checks which determine if a goto label
exists and is in the correct function scope. The assembler will complain
about gotos to labels that don't exist, but the current implementation
allows for jumps between functions, which is not according to the C standard.
Fixing this would require two passes, which is annoying.